Fixed the ddG API class to have a default username so as not to break existing checkouts with the addition of this parameter.
Added a test function for my graph (which I'm calling an abacus graph). I'll use this to fix the matplotlib global variables issue and to tune the graph scale etc.

Added a cached_results parameter to analyze_results. This allows you to pass in cached results and avoids having to run a database query each time which should speed up debugging. I also added a num_datapoints parameter for debugging.