Adding extra function parameters to match the updated database schema.
Fixing a bug in the UniProt mapping when adding PDB files which were derived from RCSB file (their ID cannot be used to look up the corresponding UniProt IDs - the original PDB's ID should be used).

Adding an option to allow the PDBMoleculeChain entry to fail. There is a practical use - if we take a PDB file with 8 chains and strip these down to 2 but keep the headers then the PDB import will fail because there is no PDBChain record associated with the molecule data in the header.