The reason why checkpointing failed to improve the times was that I removed the code to create the checkpointing file...
We still need to call get_job_files as Kyle uses the metadata to populate the job dict so the old approach of not querying the database would not work anyway.
We now tell the API to query the database but ask for a file size of zero which does not access the viewonly/deferred fields in the ORM (SQLAlchemy win!).
The time for checkpointed cases is 10% of the old time so this is a nice speedup even if it will rarely be used.