Updating the PPMutagenesis table (new field added).
Making changes to the DeltaE tables - we only need two tables to store the data. The API will have to be a little
more clever regarding wildtype measurements but I guess that it just sucks to be the API; this design is a lot cleaner.

Note that PPIDataSetDE uses PPComplexID and SetNumber. PPIDataSetDDG should be updated to also use these fields per
ticket:1457 as we otherwise introduce ambiguities when we have a choices of PDB set for a complex.