Improving add_complex to handle the case where a complex or a very similar complex already exists in the database. Before, adding the exact same complex (e.g. when running a data import script) would result in exceptions. As part of this, I am adding the private functions lookup_pdb_set and lookup_complex_by_details. These improve the code to search for complex definitions in the database.
Adding the private function get_bound_pdb_set_details which returns the pdb_id and complex partner definitions
(left PDB chains, right PDB chains) for complexes where all chains share the same PDB structure.
This function is used in the work-in-progress add_ssm_dataset function which makes it easier to add SSM datasets to the database.

Started work on the add_ssm_dataset function which is based on the import script I used for the Psd95-CRIPT SSM dataset import.